Move PrepareChargeVersionService
